Separation
Have view more sane
