+++ src/message.rs
@@ -1,0 +1,9 @@
+#[derive(Debug, Clone)]
+pub enum Message {
+    None,
+    Expand,
+    LoadFile,
+    Filter(String),
+    Submitted,
+    WindowResized(u32, u32),
+}
